requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
pydantic-core==2.16.3
pandas
aiohttp
aiosmtplib
playwright
openai
python-dotenv
motor
python-multipart
websockets
beautifulsoup4
bson
a2wsgi